Move fixture to conftest.py
