+++ tests/plots/test_violin.py
@@ -1,7 +1,6 @@
 import matplotlib.pyplot as plt
 import pytest
 import shap
-from .utils import explainer # (pytest fixture do not remove) pylint: disable=unused-import
 
 
 @pytest.mark.mpl_image_compare
